feat!: adopt web-identity-schemas, drop zod v3
Make `web-identity-schemas` (v0.3.0) the source of truth for DID/JWT/VC
validation schemas and DID/JWT types, and drop Zod v3 support.

- Collapse the `./schemas/zod/v3` + `./schemas/zod/v4` subpath exports into a
  single `./schemas/zod` (Zod v4) across every package (and `./a2a/schemas/zod`);
  bump the `zod` optional peer to `^4.0.0` and the catalog to `4.4.3`.
- `did`: `didUriSchema`/`isDidUri` now delegate to w-i-s `DidSchema`/`isDid`
  (full DID-core validation); `DidUri` re-exported from w-i-s `Did`.
- `jwt`: `jwtPayloadSchema` backed by w-i-s; `JwtString` re-exported from w-i-s;
  `jwtStringSchema` keeps ACK's non-empty-signature regex and `jwtHeaderSchema`
  keeps the narrow ES256/ES256K/EdDSA alg set.
- `vc`: `credentialSchema` backed by w-i-s `CredentialV1Schema`, preserving the
  issuer string->{id} normalization and `JwtProof2020` passthrough; ACK's VC
  types stay local (did-jwt-vc remains the signing engine).
- CAIP, payment, A2A, controller-claim, and JwtProof2020 schemas remain
  hand-rolled. Add `web-identity-schemas` as a dependency of did/jwt/vc.
- Update issuer route tests for the new DID validation error shape; update READMEs,
  AGENTS.md, and add a changeset.

AI usage: implemented by Claude Code (plan hardened via a multi-agent panel-plan
loop). See AI_POLICY.md.
