+++ packages/ack-id/package.json
@@ -33,16 +33,8 @@
       "default": "./dist/schemas/valibot.js"
     },
     "./schemas/zod": {
-      "types": "./dist/schemas/zod/v3.d.ts",
-      "default": "./dist/schemas/zod/v3.js"
-    },
-    "./schemas/zod/v3": {
-      "types": "./dist/schemas/zod/v3.d.ts",
-      "default": "./dist/schemas/zod/v3.js"
-    },
-    "./schemas/zod/v4": {
-      "types": "./dist/schemas/zod/v4.d.ts",
-      "default": "./dist/schemas/zod/v4.js"
+      "types": "./dist/schemas/zod.d.ts",
+      "default": "./dist/schemas/zod.js"
     },
     "./a2a": {
       "types": "./dist/a2a/index.d.ts",
@@ -53,16 +45,8 @@
       "default": "./dist/a2a/schemas/valibot.js"
     },
     "./a2a/schemas/zod": {
-      "types": "./dist/a2a/schemas/zod/v3.d.ts",
-      "default": "./dist/a2a/schemas/zod/v3.js"
-    },
-    "./a2a/schemas/zod/v3": {
-      "types": "./dist/a2a/schemas/zod/v3.d.ts",
-      "default": "./dist/a2a/schemas/zod/v3.js"
-    },
-    "./a2a/schemas/zod/v4": {
-      "types": "./dist/a2a/schemas/zod/v4.d.ts",
-      "default": "./dist/a2a/schemas/zod/v4.js"
+      "types": "./dist/a2a/schemas/zod.d.ts",
+      "default": "./dist/a2a/schemas/zod.js"
     }
   },
   "publishConfig": {
@@ -91,7 +75,7 @@
   },
   "peerDependencies": {
     "@a2a-js/sdk": "^0.2.2",
-    "zod": "^3.25.0"
+    "zod": "^4.0.0"
   },
   "peerDependenciesMeta": {
     "@a2a-js/sdk": {
