+++ packages/ack-id/src/schemas/zod.ts
@@ -1,4 +1,4 @@
-import { z } from "zod/v3"
+import * as z from "zod"
 
 export const controllerClaimSchema = z.object({
   id: z.string(),
