--- packages/ack-id/src/schemas/zod/v4.ts
@@ -1,6 +1,0 @@
-import * as z from "zod/v4"
-
-export const controllerClaimSchema = z.object({
-  id: z.string(),
-  controller: z.string(),
-})
