+++ packages/agentcommercekit/src/a2a/schemas/zod.ts
@@ -1,0 +1,1 @@
+export * from "@agentcommercekit/ack-id/a2a/schemas/zod"
