--- packages/agentcommercekit/src/a2a/schemas/zod/v3.ts
@@ -1,1 +1,0 @@
-export * from "@agentcommercekit/ack-id/a2a/schemas/zod/v3"
