--- packages/agentcommercekit/src/schemas/zod/v4.ts
@@ -1,6 +1,0 @@
-export * from "@agentcommercekit/ack-pay/schemas/zod/v4"
-export * from "@agentcommercekit/ack-id/schemas/zod/v4"
-export * from "@agentcommercekit/caip/schemas/zod/v4"
-export * from "@agentcommercekit/did/schemas/zod/v4"
-export * from "@agentcommercekit/jwt/schemas/zod/v4"
-export * from "@agentcommercekit/vc/schemas/zod/v4"
