+++ packages/jwt/README.md
@@ -57,9 +57,8 @@ The package provides schemas for validating JWT strings with Zod and Valibot:
 
 // Valibot
 import { jwtStringSchema } from "@agentcommercekit/jwt/schemas/valibot"
-// Zod v3
-import { jwtStringSchema } from "@agentcommercekit/jwt/schemas/zod/v3"
-import { jwtStringSchema } from "@agentcommercekit/jwt/schemas/zod/v4"
+// Zod schema
+import { jwtStringSchema } from "@agentcommercekit/jwt/schemas/zod"
 ```
 
 ## API
