+++ packages/jwt/package.json
@@ -29,16 +29,8 @@
       "default": "./dist/index.js"
     },
     "./schemas/zod": {
-      "types": "./dist/schemas/zod/v3.d.ts",
-      "default": "./dist/schemas/zod/v3.js"
-    },
-    "./schemas/zod/v3": {
-      "types": "./dist/schemas/zod/v3.d.ts",
-      "default": "./dist/schemas/zod/v3.js"
-    },
-    "./schemas/zod/v4": {
-      "types": "./dist/schemas/zod/v4.d.ts",
-      "default": "./dist/schemas/zod/v4.js"
+      "types": "./dist/schemas/zod.d.ts",
+      "default": "./dist/schemas/zod.js"
     },
     "./schemas/valibot": {
       "types": "./dist/schemas/valibot.d.ts",
@@ -57,7 +49,8 @@
   },
   "dependencies": {
     "@agentcommercekit/keys": "workspace:*",
-    "did-jwt": "8.0.18"
+    "did-jwt": "8.0.18",
+    "web-identity-schemas": "catalog:"
   },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
@@ -66,7 +59,7 @@
   },
   "peerDependencies": {
     "valibot": "^1.0.0",
-    "zod": "^3.25.0"
+    "zod": "^4.0.0"
   },
   "peerDependenciesMeta": {
     "valibot": {
