+++ packages/jwt/src/schemas/valibot.ts
@@ -1,21 +1,22 @@
 import * as v from "valibot"
+import { JwtPayloadSchema } from "web-identity-schemas/valibot"
 
-import type { JwtHeader, JwtPayload } from "../create-jwt"
+import type { JwtHeader } from "../create-jwt"
 import { jwtAlgorithms } from "../jwt-algorithm"
-import type { JwtString } from "../jwt-string"
 
-export const jwtPayloadSchema = v.pipe(
-  v.looseObject({
-    iss: v.optional(v.string()),
-    sub: v.optional(v.string()),
-    aud: v.optional(v.union([v.string(), v.array(v.string())])),
-    iat: v.optional(v.number()),
-    nbf: v.optional(v.number()),
-    exp: v.optional(v.number()),
-  }),
-  v.custom<JwtPayload>(() => true),
-)
+/**
+ * JWT payload schema, backed by `web-identity-schemas`. Kept as a raw
+ * loose-object schema so consumers can spread `.entries`.
+ */
+export const jwtPayloadSchema = JwtPayloadSchema
 
+/**
+ * JWT header restricted to the algorithms ACK signs with.
+ *
+ * `web-identity-schemas`' `JwtHeaderSignedSchema` allows all JOSE algorithms;
+ * ACK only ever signs with `ES256`/`ES256K`/`EdDSA`, so this wrapper keeps the
+ * narrow set.
+ */
 export const jwtHeaderSchema = v.pipe(
   v.looseObject({
     typ: v.literal("JWT"),
@@ -24,8 +25,13 @@ export const jwtHeaderSchema = v.pipe(
   v.custom<JwtHeader>(() => true),
 )
 
+/**
+ * JWT string schema requiring a non-empty signature segment.
+ *
+ * Stricter than `web-identity-schemas`' `JwtStringSchema` (which permits an
+ * empty signature) — see {@link isJwtString}.
+ */
 export const jwtStringSchema = v.pipe(
   v.string(),
   v.regex(/^[a-zA-Z0-9-_]+\.[a-zA-Z0-9-_]+\.[a-zA-Z0-9-_]+$/),
-  v.transform((input) => input as JwtString),
 )
