--- packages/jwt/src/schemas/zod/v4.ts
@@ -1,30 +1,0 @@
-import * as z from "zod/v4"
-
-import type { JwtHeader, JwtPayload } from "../../create-jwt"
-import { jwtAlgorithms } from "../../jwt-algorithm"
-import { isJwtString } from "../../jwt-string"
-
-export const jwtPayloadSchema = z
-  .looseObject({
-    iss: z.optional(z.string()),
-    sub: z.optional(z.string()),
-    aud: z.optional(z.union([z.string(), z.array(z.string())])),
-    iat: z.optional(z.number()),
-    nbf: z.optional(z.number()),
-    exp: z.optional(z.number()),
-  })
-  .refine((val): val is JwtPayload => true)
-
-export const jwtHeaderSchema = z
-  .looseObject({
-    typ: z.literal("JWT"),
-    alg: z.enum(jwtAlgorithms),
-  })
-  .refine((val): val is JwtHeader => true)
-
-export const jwtStringSchema = z
-  .string()
-  .regex(/^[a-zA-Z0-9-_]+\.[a-zA-Z0-9-_]+\.[a-zA-Z0-9-_]+$/)
-  .refine((input) => isJwtString(input), {
-    message: "Invalid JWT string",
-  })
