+++ packages/vc/package.json
@@ -29,16 +29,8 @@
       "default": "./dist/index.js"
     },
     "./schemas/zod": {
-      "types": "./dist/schemas/zod/v3.d.ts",
-      "default": "./dist/schemas/zod/v3.js"
-    },
-    "./schemas/zod/v3": {
-      "types": "./dist/schemas/zod/v3.d.ts",
-      "default": "./dist/schemas/zod/v3.js"
-    },
-    "./schemas/zod/v4": {
-      "types": "./dist/schemas/zod/v4.d.ts",
-      "default": "./dist/schemas/zod/v4.js"
+      "types": "./dist/schemas/zod.d.ts",
+      "default": "./dist/schemas/zod.js"
     },
     "./schemas/valibot": {
       "types": "./dist/schemas/valibot.d.ts",
@@ -61,14 +53,15 @@
     "@agentcommercekit/keys": "workspace:*",
     "bit-buffers": "catalog:",
     "did-jwt-vc": "4.0.16",
-    "valibot": "catalog:"
+    "valibot": "catalog:",
+    "web-identity-schemas": "catalog:"
   },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
     "zod": "catalog:"
   },
   "peerDependencies": {
-    "zod": "^3.25.0"
+    "zod": "^4.0.0"
   },
   "peerDependenciesMeta": {
     "zod": {
